refactor(api-file-manager-server): split upload routes into separate files
One route per file: UploadSingleFileRoute.ts, UploadPartRoute.ts.
Shared helpers extracted to routes/utils.ts.
